refactor: - migrating to streaming
