+++ backend/server/cmd/router.go
@@ -16,7 +16,6 @@ import (
 	"github.com/mordilloSan/LinuxIO/backend/server/auth"
 	"github.com/mordilloSan/LinuxIO/backend/server/benchmark"
 	"github.com/mordilloSan/LinuxIO/backend/server/bridge/handlers/config"
-	"github.com/mordilloSan/LinuxIO/backend/server/bridge/handlers/control"
 	"github.com/mordilloSan/LinuxIO/backend/server/bridge/handlers/filebrowser"
 	"github.com/mordilloSan/LinuxIO/backend/server/web"
 )
@@ -46,7 +45,6 @@ func BuildRouter(cfg Config, sm *session.Manager) *gin.Engine {
 	// --- API routes ---
 	auth.RegisterAuthRoutes(r, sm, cfg.Env, cfg.Verbose, cfg.BridgeBinaryOverride)
 	config.RegisterThemeRoutes(r.Group("/theme", sm.RequireSession()))
-	control.RegisterControlRoutes(r.Group("/control", sm.RequireSession()))
 	filebrowser.RegisterFilebrowserRoutes(r.Group("/navigator", sm.RequireSession()))
 
 	// --- WebSocket ---
